+++ lsql-core/src/main/java/com/w11k/lsql/LSql.java
@@ -6,6 +6,7 @@ import com.google.common.base.CaseFormat;
 import com.google.common.base.Optional;
 import com.google.common.base.Splitter;
 import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.w11k.lsql.converter.Converter;
 import com.w11k.lsql.dialects.GenericDialect;
@@ -23,6 +24,7 @@ import org.slf4j.LoggerFactory;
 import javax.sql.DataSource;
 import java.sql.*;
 import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Callable;
 
@@ -194,10 +196,15 @@ public class LSql {
 
         ResultSet tables = md.getTables(null, null, null, new String[]{"TABLE"});
         try {
+            List<String> foundTables = Lists.newLinkedList();
             while (tables.next()) {
                 String sqlTableName = tables.getString(3);
                 String javaTableName = identifierSqlToJava(sqlTableName);
-                Table table = table(javaTableName);
+                foundTables.add(javaTableName);
+            }
+
+            for (String foundTable : foundTables) {
+                Table table = table(foundTable);
                 this.logger.debug("Reading table " + table.getSchemaAndTableName());
             }
         } catch (Exception e) {